requirements.txt
Django==1.11.3
django-simple-history==1.9.0
pkg-resources==0.0.0
pytz==2017.2
